docs/use-cases/actuarial-analysis-solution/pyproject.toml
[project]
name = "actuarial-analysis-mcp"
version = "0.1.0"
description = "Actuarial Analysis Solution MCP integration with QuickSuite"
requires-python = ">=3.11"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.0.0",
]

[tool.uv]
dev-dependencies = []
